+++ okdata/sdk/webhook/client.py
@@ -11,10 +11,8 @@ class WebhookClient(SDK):
         super().__init__(config, auth, env)
         self.api_url = f"{self.config.get('permissionApiUrl')}/webhooks"
 
-    def create_webhook_token(
-        self, dataset_id: str, service_name: str, retries: int = 0
-    ):
-        """Create a webhook token associated with a dataset and an okdata service.
+    def create_webhook_token(self, dataset_id: str, operation: str, retries: int = 0):
+        """Create a webhook token for a given operation on a dataset.
 
         Return a dictionary on the form:
 
@@ -22,18 +20,14 @@ class WebhookClient(SDK):
             "token": "774d8f35-fb4b-4e06-9d7f-54d1a08589b4",
             "created_by": "your-username",
             "dataset_id": "some-dataset",
-            "service": "service-account-some-service",
+            "operation": "write",
             "created_at": "2020-02-29T00:00:00+00:00",
             "expires_at": "2022-02-28T00:00:00+00:00",
             "is_active": True,
           }
-        Note!
-        The service_name parameter should NOT be prefixed with "service-account-".
         """
         url = f"{self.api_url}/{dataset_id}/tokens"
-        request_body = {"service": service_name}
-
-        return self.post(url, request_body, retries=retries).json()
+        return self.post(url, {"operation": operation}, retries=retries).json()
 
     def list_webhook_tokens(self, dataset_id: str, retries: int = 0):
         """List all webhook tokens associated with a dataset.
@@ -44,13 +38,12 @@ class WebhookClient(SDK):
               "token": "774d8f35-fb4b-4e06-9d7f-54d1a08589b4",
               "created_by": "your-username",
               "dataset_id": "some-dataset",
-              "service": "service-account-some-service",
+              "operation": "write",
               "created_at": "2020-02-29T00:00:00+00:00",
               "expires_at": "2022-02-28T00:00:00+00:00",
               "is_active": True,
             },
-          .
-          .
+            ⋮
           ]
         """
         url = f"{self.api_url}/{dataset_id}/tokens"
@@ -67,8 +60,10 @@ class WebhookClient(SDK):
         url = f"{self.api_url}/{dataset_id}/tokens/{token}"
         return self.delete(url, retries=retries).json()
 
-    def authorize_webhook_token(self, dataset_id: str, token: str, retries: int = 0):
-        """Check if a webhook token has access to perform operations on a dataset from a logged in service.
+    def authorize_webhook_token(
+        self, dataset_id: str, token: str, operation: str, retries: int = 0
+    ):
+        """Check if a webhook token has access to an operation on a dataset.
 
         Return a dictionary on the form:
           {
@@ -76,5 +71,5 @@ class WebhookClient(SDK):
             "reason": None
           }
         """
-        url = f"{self.api_url}/{dataset_id}/tokens/{token}/authorize"
+        url = f"{self.api_url}/{dataset_id}/tokens/{token}/authorize?operation={operation}"
         return self.get(url, retries=retries).json()